Cb extract validator (#666)
