+++ clinicadl/commandline/pipelines/interpret/options.py
@@ -2,28 +2,28 @@ import click
 
 from clinicadl.config.config_utils import get_default_from_config_class as get_default
 from clinicadl.config.config_utils import get_type_from_config_class as get_type
-from clinicadl.interpret.config import InterpretConfig
+from clinicadl.interpret.config import InterpretBaseConfig
 
 # interpret specific
 name = click.argument(
     "name",
-    type=get_type("name", InterpretConfig),
+    type=get_type("name", InterpretBaseConfig),
 )
 method = click.argument(
     "method",
-    type=get_type("method", InterpretConfig),  # ["gradients", "grad-cam"]
+    type=get_type("method", InterpretBaseConfig),  # ["gradients", "grad-cam"]
 )
 level = click.option(
     "--level_grad_cam",
-    type=get_type("level", InterpretConfig),
-    default=get_default("level", InterpretConfig),
+    type=get_type("level", InterpretBaseConfig),
+    default=get_default("level", InterpretBaseConfig),
     help="level of the feature map (after the layer corresponding to the number) chosen for grad-cam.",
     show_default=True,
 )
 target_node = click.option(
     "--target_node",
-    type=get_type("target_node", InterpretConfig),
-    default=get_default("target_node", InterpretConfig),
+    type=get_type("target_node", InterpretBaseConfig),
+    default=get_default("target_node", InterpretBaseConfig),
     help="Which target node the gradients explain. Default takes the first output node.",
     show_default=True,
 )
