+++ clinicadl/commandline/pipelines/predict/options.py
@@ -1,13 +1,11 @@
 import click
 
 from clinicadl.config.config_utils import get_default_from_config_class as get_default
-from clinicadl.predict.config import PredictConfig
+from clinicadl.predictor.config import PredictConfig
 
 # predict specific
 use_labels = click.option(
     "--use_labels/--no_labels",
-    show_default=True,
-    default=get_default("use_labels", PredictConfig),
     help="Set this option to --no_labels if your dataset does not contain ground truth labels.",
 )
 save_tensor = click.option(
