+++ clinicadl/commandline/pipelines/train/reconstruction/cli.py
@@ -13,7 +13,6 @@ from clinicadl.commandline.modules_options import (
     optimizer,
     reproducibility,
     split,
-    ssda,
     transforms,
     validation,
 )
@@ -63,12 +62,6 @@ from clinicadl.utils.iotools.train_utils import merge_cli_and_config_file_option
 @dataloader.batch_size
 @dataloader.sampler
 @dataloader.n_proc
-# ssda option
-@ssda.ssda_network
-@ssda.caps_target
-@ssda.tsv_target_lab
-@ssda.tsv_target_unlab
-@ssda.preprocessing_json_target
 # Cross validation
 @split.n_splits
 @split.split
