+++ clinicadl/commandline/pipelines/train/resume/cli.py
@@ -16,4 +16,4 @@ def cli(input_maps_directory, split):
     INPUT_MAPS_DIRECTORY is the path to the MAPS folder where training job has started.
     """
     trainer = Trainer.from_maps(input_maps_directory)
-    trainer.resume(split)
+    trainer.resume()
