+++ clinicadl/resources/config/train_config.toml
@@ -4,7 +4,6 @@
 [Model]
 architecture = "default" # ex : Conv5_FC3
 multi_network = false
-ssda_network = false
 
 [Architecture]
 # CNN
