+++ clinicadl/utils/cli_param/option.py
@@ -58,13 +58,6 @@ variables_of_interest = click.option(
     multiple=True,
     default=None,
 )
-ssda_network = click.option(
-    "--ssda_network",
-    type=bool,
-    default=False,
-    show_default=True,
-    help="ssda training.",
-)
 valid_longitudinal = click.option(
     "--valid_longitudinal/--valid_baseline",
     type=bool,
